+++ COT/tests/test_doctests.py
@@ -40,4 +40,5 @@ def load_tests(*_):
     suite.addTests(DocTestSuite('COT.edit_properties'))
     suite.addTests(DocTestSuite('COT.file_reference'))
     suite.addTests(DocTestSuite('COT.platforms'))
+    suite.addTests(DocTestSuite('COT.ui_shared'))
     return suite