+++ COT/ui_shared.py
@@ -3,7 +3,7 @@
 # ui_shared.py - abstraction between CLI and GUI
 #
 # December 2014, Glenn F. Matthews
-# Copyright (c) 2014-2016 the COT project developers.
+# Copyright (c) 2014-2017 the COT project developers.
 # See the COPYRIGHT.txt file at the top-level directory of this distribution
 # and at https://github.com/glennmatthews/cot/blob/master/COPYRIGHT.txt.
 #
@@ -14,10 +14,27 @@
 # of COT, including this file, may be copied, modified, propagated, or
 # distributed except according to the terms contained in the LICENSE.txt file.
 
-"""Abstract user interface superclass."""
+"""Generic user interface functions and abstract user interface superclass.
+
+**Functions**
+
+.. autosummary::
+  :nosignatures:
+
+  pretty_bytes
+  to_string
+
+**Classes**
+
+.. autosummary::
+  :nosignatures:
+
+  UI
+"""
 
 import logging
 import sys
+import xml.etree.ElementTree as ET
 
 from verboselogs import VerboseLogger
 
@@ -25,6 +42,99 @@ from verboselogs import VerboseLogger
 # This lets us be a bit more fine-grained in our logging verbosity.
 logging.setLoggerClass(VerboseLogger)
 logger = logging.getLogger(__name__)
+
+
+def to_string(obj):
+    """Get string representation of an object, special-case for XML Element.
+
+    Args:
+      obj (object): Object to represent as a string.
+    Returns:
+      str: string representation
+    Examples:
+      ::
+
+        >>> to_string("Hello")
+        'Hello'
+        >>> to_string(27.5)
+        '27.5'
+        >>> e = ET.Element('hello', attrib={'key': 'value'})
+        >>> print(e)   # doctest: +ELLIPSIS
+        <Element ...hello... at ...>
+        >>> print(to_string(e))
+        <hello key="value" />
+    """
+    if ET.iselement(obj):
+        if sys.version_info[0] >= 3:
+            return ET.tostring(obj, encoding='unicode')
+        else:
+            return ET.tostring(obj)
+    else:
+        return str(obj)
+
+
+def pretty_bytes(byte_value, base_shift=0):
+    """Pretty-print the given bytes value.
+
+    Args:
+      byte_value (float): Value
+      base_shift (int): Base value of byte_value
+            (0 = bytes, 1 = KiB, 2 = MiB, etc.)
+
+    Returns:
+      str: Pretty-printed byte string such as "1.00 GiB"
+
+    Examples:
+      ::
+
+        >>> pretty_bytes(512)
+        '512 B'
+        >>> pretty_bytes(512, 2)
+        '512 MiB'
+        >>> pretty_bytes(65536, 2)
+        '64 GiB'
+        >>> pretty_bytes(65547)
+        '64.01 KiB'
+        >>> pretty_bytes(65530, 3)
+        '63.99 TiB'
+        >>> pretty_bytes(1023850)
+        '999.9 KiB'
+        >>> pretty_bytes(1024000)
+        '1000 KiB'
+        >>> pretty_bytes(1048575)
+        '1024 KiB'
+        >>> pretty_bytes(1049200)
+        '1.001 MiB'
+        >>> pretty_bytes(2560)
+        '2.5 KiB'
+        >>> pretty_bytes(.0001, 3)
+        '104.9 KiB'
+        >>> pretty_bytes(.01, 1)
+        '10 B'
+        >>> pretty_bytes(.001, 1)
+        '1 B'
+        >>> pretty_bytes(.0001, 1)
+        '0 B'
+        >>> pretty_bytes(100, -1)
+        Traceback (most recent call last):
+            ...
+        ValueError: base_shift must not be negative
+    """
+    if base_shift < 0:
+        raise ValueError("base_shift must not be negative")
+    tags = ["B", "KiB", "MiB", "GiB", "TiB"]
+    byte_value = float(byte_value)
+    shift = base_shift
+    while byte_value >= 1024.0:
+        byte_value /= 1024.0
+        shift += 1
+    while byte_value < 1.0 and shift > 0:
+        byte_value *= 1024.0
+        shift -= 1
+    # Fractions of a byte should be considered a rounding error:
+    if shift == 0:
+        byte_value = round(byte_value)
+    return "{0:.4g} {1}".format(byte_value, tags[shift])
 
 
 class UI(object):
@@ -188,3 +298,8 @@ class UI(object):
           NotImplementedError: Must be implemented by a subclass.
         """
         raise NotImplementedError("No implementation of get_password()")
+
+
+if __name__ == "__main__":
+    import doctest
+    doctest.testmod()